+++ src/main/scala/org/izolotov/crawler/parser/TramontanaParser.scala
@@ -17,11 +17,11 @@ object TramontanaParser extends Parser[Product] {
       val brand = doc.select("#product-all-brands > span").first.text()
       val title = doc.select("h1.text-uppercase").first.text()
       val category = doc.select(".breadcrumb-item").asScala.map(e => e.text())
-      val oldPrice: Option[Int] = doc.select("#product-old-price").text() match {
+      val oldPrice: Option[Float] = doc.select("#product-old-price").text() match {
         case "" => None
         case price => Some(Util.parsePrice(price))
       }
-      val price: Int = Util.parsePrice(doc.select("#product-price").first.text())
+      val price: Float = Util.parsePrice(doc.select("#product-price").first.text())
       Product(url, StoreName, brand, title, category, price, oldPrice, Currency.Rub.toString)
     } catch {
       case exc => Product(url = url, store = StoreName, parseError = Some(exc.toString))
